requirements.txt
linecache2
cx_Oracle
fastapi
hypercorn
pydantic
uvicorn
requests
Pillow